Modify : users app models.py

+++ users/models.py
@@ -4,7 +4,7 @@ from core.models import TimeStamp
 class User(TimeStamp):
     name              = models.CharField(max_length=50, null=True)
     nickname          = models.CharField(max_length=50, null=True)
-    kakao_id          = models.IntegerField()
+    kakao_id          = models.BigIntegerField()
     email             = models.CharField(max_length=100, null=True)
     description       = models.CharField(max_length=500, null=True)
     profile_image_url = models.CharField(max_length=2000, default='https://cdn.pixabay.com/photo/2016/03/31/19/58/avatar-1295429_960_720.png')
